backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic[email]==2.8.0
phonenumbers==8.13.40
SQLAlchemy==2.0.31
python-dotenv==1.0.1
alembic==1.13.2
psycopg2==2.9.9
fastapi-pagination==0.12.26